Moved cartList and CartItem to commons and added hidden price cartList option
